backend/requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0

# ── Database ──────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ── Cache ─────────────────────────────────────────────────────────────────
redis==5.0.4

# ── Storage ───────────────────────────────────────────────────────────────
minio==7.2.7

# ── Vector DB ─────────────────────────────────────────────────────────────
chromadb==0.5.0

# ── AI / Agents ───────────────────────────────────────────────────────────
google-generativeai==0.8.3
langchain==0.2.3
langgraph==0.1.1

# ── Legacy (kept for fallback) ────────────────────────────────────────────
anthropic==0.28.0
langchain-anthropic==0.1.15

# ── PDF Extraction (lightweight) ──────────────────────────────────────────
pdfplumber==0.11.0
Pillow==10.3.0

# ── Document Output ───────────────────────────────────────────────────────
python-docx==1.1.2

# ── Research ──────────────────────────────────────────────────────────────
tavily-python==0.3.3

# ── Graph ─────────────────────────────────────────────────────────────────
networkx==3.3

# ── REMOVED (too heavy for MVP) ───────────────────────────────────────────
# torch, transformers, sentence-transformers → Claude handles all NLP
# opencv-python-headless, camelot-py[cv]     → pdfplumber replaces camelot
# scikit-learn, numpy, pandas                → inline math replaces sklearn
# pytesseract, pdf2image                     → OCR skipped for MVP
# weasyprint                                 → python-docx sufficient


